Create 0xbfc4dc47c254acce7fc735de552b8b0fa15a0453
cowrie ocean a crypto currency for African ecommerce and social media earnings,  in Africa cowries is not just a currency but a symbol of affluence, a spiritual token that has a cross dimensional value, it also features as a dynamic fashion ornament. Cowries as a digital currency is uniting the values and extending the reach. Rebranding the new era Africa and updating our ancient cultural values